Made the cop dynamic for other methods
For all the methods that follow that pattern, this adds a fix.

This also adds in a spacing fix so autocorrect doesnt force another correction by the line-indentation cop.

I am thinking though that the `return "#{replaced_text}, #{optional_param.source}" if optional_param` line might be problematic when we use parameters. I am thinking that we might want to put a placeholder always and replace it with the parameter when appropriate and an empty strring when not. Or some other change.